all compose opps working

+++ app/lib/envfile.server.ts
@@ -1,14 +1,7 @@
 import env from "./env.server"
 import fs from 'fs/promises'
 import path from 'path'
-import { emitter } from "./emitter.server"
-
-export async function getComposeFiles() {
-  const configFolderENV = await readEnvFile()
-  const separator = configFolderENV.COMPOSE_PATH_SEPARATOR || ':'
-  const composeFiles = (configFolderENV.COMPOSE_FILE || 'docker-compose.yml').split(separator)
-  return composeFiles
-}
+import fileExists from "./fileExists"
 
 export function parseEnvFileText(envText: string) {
   const envVars = {} as Record<string, string>
@@ -21,17 +14,25 @@ export function parseEnvFileText(envText: string) {
   return envVars
 }
 
-export async function readEnvFile() {
-  const text = await fs.readFile(path.join(env.configFolder, '.env'), {
-    encoding: 'utf8',
-    flag: 'a+',
-  })
+export async function readEnvFile(projectFolder: string) {
+  const text = await fs.readFile(
+    path.join(env.configFolder, projectFolder, '.env'),
+    { encoding: 'utf8' }
+  )
   return parseEnvFileText(text)
 }
 
-export async function addToDotEnv(filename: string) {
-  const configFolderENV = await readEnvFile()
-  const separator = configFolderENV.COMPOSE_PATH_SEPARATOR || ':'
+const COMPOSE_DEFAULT_SEPARATOR = ':' as const
+
+export async function addToDotEnv(projectFolder: string, filename: string) {
+  const envFile = path.join(env.configFolder, projectFolder, '.env')
+  const envExists = await fileExists(envFile)
+  if (!envExists) {
+    return
+  }
+
+  const configFolderENV = await readEnvFile(projectFolder)
+  const separator = configFolderENV.COMPOSE_PATH_SEPARATOR || COMPOSE_DEFAULT_SEPARATOR
   const fileListText = configFolderENV.COMPOSE_FILE || ''
   const composeFiles = new Set(fileListText.split(separator))
   composeFiles.add(filename)
@@ -44,13 +45,21 @@ export async function addToDotEnv(filename: string) {
     .map(([key, value]) => `${key}=${value}`)
     .join('\n')
 
-  await fs.writeFile(path.join(env.configFolder, '.env'), newDotEnv)
-  emitter.emit('message', `Added ${filename} to .env`)
+  await fs.writeFile(
+    path.join(env.configFolder, projectFolder, '.env'),
+    newDotEnv
+  )
 }
 
-export async function removeFromDotEnv(filename: string) {
-  const configFolderENV = await readEnvFile()
-  const separator = configFolderENV.COMPOSE_PATH_SEPARATOR || ':'
+export async function removeFromDotEnv(projectFolder: string, filename: string) {
+  const envFile = path.join(env.configFolder, projectFolder, '.env')
+  const envExists = await fileExists(envFile)
+  if (!envExists) {
+    return
+  }
+
+  const configFolderENV = await readEnvFile(projectFolder)
+  const separator = configFolderENV.COMPOSE_PATH_SEPARATOR || COMPOSE_DEFAULT_SEPARATOR
   const fileListText = configFolderENV.COMPOSE_FILE || ''
   const composeFiles = new Set(fileListText.split(separator))
   composeFiles.delete(filename)
@@ -63,6 +72,8 @@ export async function removeFromDotEnv(filename: string) {
     .map(([key, value]) => `${key}=${value}`)
     .join('\n')
 
-  await fs.writeFile(path.join(env.configFolder, '.env'), newDotEnv)
-  emitter.emit('message', `Removed ${filename} from .env`)
+  await fs.writeFile(
+    path.join(env.configFolder, projectFolder, '.env'),
+    newDotEnv
+  )
 }
